do_vagrant_cbs.sh should wait for CBS tasks to complete, when in a pipe
When the output of do_vagrant_cbs.sh is redirected, koji should not
return before the CBS tasks are complete.  Otherwise, the following
process in the pipe will get incomplete information about the
still-running children tasks, if it tries to use "cbs taskinfo -r".
This is used by the build on ci.centos.org to get the full path of the
generated Vagrant images.