Disable kdump
kdump is something seldom needed in a VM (if ever), and it needs to
reserve 160MB + 2bits/4kB RAM. The automatic memory allocation is
disabled for systems with less than 2GB RAM, so it didn't work on most
Vagrant boxes anyway. Only disabled for centos/7, since CentOS Linux 6
asks if it should enable kdump during firstboot (which is disabled in
centos/6).