Do not tell the kernel to use a serial console
VirtualBox and Hyper-V do not have a serial console by default. This is
mostly harmless when the console is only used for output, but relabeling
the filesystem insists on having a tty, passing init=/bin/bash or
booting a fully sealed template also require a tty - resulting in a
system that always hangs indefinitely on boot (or, in the init case,
produces a kernel panic).