Use ttyS0 as the default system console
If we use tty0 as the default, agetty starts too soon in CentOS Linux 6,
and the login prompt is followed by lots of messages from other
services.  The user has to press <enter> to get a login prompt again, in
this case.  Setting ttyS0 as the default console avoids this problem,
and we don't even need to create a /etc/init/serial-ttyS0.conf file for
upstart anymore.