Reduce image size by not installing docs (#111)
Anaconda supports omitting the docs files when installing packages. This
reduces the image size of centos/7 from 386MB to 320MB, making spinning
up a new Vagrant box faster.

It is still possible to reinstall a package with the documentation, if a
user wants that. The "--excludedocs" option of Kickstart is already used
by our Docker images; it's only to my chronic lack of time that it took
so long for the Vagrant images to use that, too.