Set the default IO scheduler to noop (#124)
The best IO scheduler for a VM is noop, not the default cfq.  We have to
send the disk sectors to be written or read as fast as possible to the
host OS, which is the only one able to optimize and reorder the disk
operations considering the needs of all hosted VMs and other
applications that might be running.

The RedHat documentation also recommends noop for guest VMs, unless they
use dedicated raw partitions (not the case for Vagrant images).