Switch the filesystem from xfs to ext4
All CentOS Linux 7 images in the last few months are unable to boot due
to apparent XFS filesystem corruption.  This is 100% reproducible with
both Image Factory and Packer when using libvirt (the images for
VirtualBox are working).  It might be caused by qemu-img, since the
initial boot performed by Packer immediately after the installation is
successful.