Use a single partition for the Vagrant images (#115)
* Use a single partition for Vagrant images

There is no performance penalty for using a swapfile instead of a swap
partition; by not defining a swap of boot partition, we allow our users
to create a swapfile themselves if they need one.

* Use Kickstart to set the bootloader timeout

* Remove several unneccessary packages

- plymouth displays a bootup screen, not needed by Vagrant
- manpages don't bring anything if we use --excludedocs
- linux-firmware isn't useful for VM hardware
- support for ext[234] and btrfs isn't needed, we use XFS

* Add the man pages again to the image

* Configure swapping to a file

We use a 2GiB file as swap, as recommended by Red Hat for systems with
up to 4G RAM.

* Add cifs-utils to the images (for Windows users)

* Switch the filesystem from xfs to ext4

All CentOS Linux 7 images in the last few months are unable to boot due
to apparent XFS filesystem corruption.  This is 100% reproducible with
both Image Factory and Packer when using libvirt (the images for
VirtualBox are working).  It might be caused by qemu-img, since the
initial boot performed by Packer immediately after the installation is
successful.